--- test-data/src/main/resources/org/openmrs/module/fhir2/api/dao/impl/FhirPerson_2_2_initial_data.xml
@@ -1,15 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<!--
-    This Source Code Form is subject to the terms of the Mozilla Public License,
-    v. 2.0. If a copy of the MPL was not distributed with this file, You can
-    obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
-    the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
-
-    Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
-    graphic logo is a trademark of OpenMRS Inc.
--->
-<dataset>
-	<person person_id="2" gender="M" birthdate="1999-12-20" dead="false" creator="1" date_created="2005-01-01 00:00:00.0" voided="false"  uuid="61b38324-e2fd-4feb-95b7-9e9a2a4400df"/>
-	<person_name person_name_id="2" preferred="true" person_id="2" given_name="John" middle_name="F" family_name="Doe" creator="1" date_created="2005-01-01 00:00:00.0" voided="false" uuid="7e2acadc-5073-4a39-914a-debcbec8c1c9"/>
-	<person_address person_address_id="2" preferred="true" person_id="2" city_village="Johnson City" state_province="TN" postal_code="37601" country="FakeB" creator="1" date_created="2005-01-01 00:00:00.0" voided="false" uuid="da5931e4-67b7-11ea-bc55-0242ac130003"/>
-</dataset>
